Make prediction list compact and add zoom effect.
The listed predictions wasted a lot screen space. Remove all padding
from them and expand the list element under the cursor, instead,
similarly to accordion. That way more addresses can be shown and the
user can still effectively click on one.

Currently it works only for desktop browsers where screen space is not
such a big issue. Touch interfaces should probably not be handled with
hover.

Also, it is annoying that the buttons below the list move when a list
element is being hovered over.